config.pyから、設定パラメーター以外をconstants.pyとpath.pyに分離。 コマンドライン処理部分を別モジュールに分離。 テンプレート関連をクラス化して、別モジュールに分離。 初期化処理をmain()から分離。
