+++ tools/yaml2rst/yaml2rst.py
@@ -1,27 +1,25 @@
 import os
 import sys
 import json
-import argparse
-import unicodedata
 import yaml
 from jsonschema import validate, ValidationError, RefResolver
-from jinja2 import Environment, FileSystemLoader
-from config import AVAILABLE_LANGUAGES, CHECK_TOOLS, SRCDIR, SCHEMA_FILENAMES, COMMON_SCHEMA_PATH, TEMPLATE_DIR, TEMPLATE_FILENAMES, MISC_INFO_SRCFILES, get_dest_dirnames, get_static_dest_files
+import initialize
+from path import SRCDIR, SCHEMA_FILENAMES, COMMON_SCHEMA_PATH, MISC_INFO_SRCFILES
+from constants import CHECK_TOOLS
 from a11y_guidelines import Category, WcagSc, InfoRef, Guideline, Check, Faq, FaqTag, CheckTool, RelationshipManager
 
 def main():
-    args = parse_args()
-    settings = process_arguments(args)
-    build_all = settings.get('build_all')
-    targets = settings.get('targets')
-    LANG = settings.get('lang')
-    DEST_DIRS = get_dest_dirnames(LANG)
-    STATIC_FILES = get_static_dest_files(LANG)
+    settings = initialize.setup_parameters()
+    build_all, targets, lang, no_check = (
+        settings.get('build_all'),
+        settings.get('targets'),
+        settings.get('lang'),
+        settings.get('no_check')
+    )
+    DEST_DIRS, STATIC_FILES, MAKEFILE_VARS = initialize.setup_constants(lang)
+    templates = initialize.setup_templates()
 
-    template_env = setup_template_environment()
-    templates = load_templates(template_env)
-
-    if not args.no_check:
+    if not no_check:
         try:
             file_content = read_file_content(COMMON_SCHEMA_PATH)
             common_schema = json.loads(file_content)
@@ -30,35 +28,7 @@ def main():
         schema_path = f'file://{SRCDIR["schema"]}/'
         resolver = RefResolver(schema_path, common_schema)
 
-    data = {
-        'guidelines': [],
-        'checks': [],
-        'faqs': [],
-        'faq_tags': [],
-        'categories': [],
-        'wcag_sc': []
-    }
-
-    makefile_vars = {
-        'all_checks_target': STATIC_FILES['all_checks'],
-        'faq_index_target': " ".join([STATIC_FILES[key] for key in ['faq_index', 'faq_tag_index', 'faq_article_index']]),
-        'wcag_mapping_target': STATIC_FILES['wcag21mapping'],
-        'priority_diff_target': STATIC_FILES['priority_diff'],
-        'miscdefs_target': STATIC_FILES['miscdefs'],
-        'wcag_sc': MISC_INFO_SRCFILES['wcag_sc'],
-        'info_src': MISC_INFO_SRCFILES['info'],
-        'gl_yaml': '',
-        'check_yaml': '',
-        'faq_yaml': ''
-    }
-    makefile_vars_list = {
-        'guideline_category_target': [],
-        'check_example_target': [],
-        'faq_article_target': [],
-        'faq_tagpage_target': [],
-        'info_to_gl_target': [],
-        'info_to_faq_target': [],
-    }
+    data, makefile_vars, makefile_vars_list = initialize.setup_variables()
 
     # Setup CheckTool instances
     for tool_id, tool_names in CHECK_TOOLS.items():
@@ -67,8 +37,8 @@ def main():
     # Set up check instances
     files = ls_dir(SRCDIR['checks'])
     makefile_vars['check_yaml'] = ' '.join(files)
-    data['checks'] = [process_yaml_file(f, SRCDIR['schema'], SCHEMA_FILENAMES['checks'], args.no_check, resolver) for f in files]
-    if not args.no_check:
+    data['checks'] = [process_yaml_file(f, SRCDIR['schema'], SCHEMA_FILENAMES['checks'], no_check, resolver) for f in files]
+    if not no_check:
         check_duplicate_values(data['checks'], 'id', 'Check ID')
     for check in data['checks']:
         Check(check)
@@ -94,8 +64,8 @@ def main():
     # Set up guideline instances
     files = ls_dir(SRCDIR['guidelines'])
     makefile_vars['gl_yaml'] = ' '.join(files)
-    data['guidelines'] = [process_yaml_file(f, SRCDIR['schema'], SCHEMA_FILENAMES['guidelines'], args.no_check, resolver) for f in files]
-    if not args.no_check:
+    data['guidelines'] = [process_yaml_file(f, SRCDIR['schema'], SCHEMA_FILENAMES['guidelines'], no_check, resolver) for f in files]
+    if not no_check:
         check_duplicate_values(data['guidelines'], 'id', 'Guideline ID')
         check_duplicate_values(data['guidelines'], 'sortKey', 'Guideline sortKey')
     for gl in data['guidelines']:
@@ -113,8 +83,8 @@ def main():
     # Set up FAQ instances
     files = ls_dir(SRCDIR['faq'])
     makefile_vars['faq_yaml'] = ' '.join(files)
-    data['faqs'] = [process_yaml_file(f, SRCDIR['schema'], SCHEMA_FILENAMES['faq'], args.no_check, resolver) for f in files]
-    if not args.no_check:
+    data['faqs'] = [process_yaml_file(f, SRCDIR['schema'], SCHEMA_FILENAMES['faq'], no_check, resolver) for f in files]
+    if not no_check:
         check_duplicate_values(data['faqs'], 'id', 'FAQ ID')
         check_duplicate_values(data['faqs'], 'sortKey', 'FAQ sortKey')
     for faq in data['faqs']:
@@ -126,47 +96,47 @@ def main():
         filename = f'{category}.rst'
         destfile = os.path.join(DEST_DIRS['guidelines'], filename)
         if build_all or destfile in targets:
-            gl_object = [gl.template_object(LANG) for gl in guidelines]
-            write_rst(templates['category_page'], {'lang': LANG, 'guidelines': gl_object}, destfile)
+            gl_object = [gl.template_object(lang) for gl in guidelines]
+            templates['category_page'].write_rst({'lang': lang, 'guidelines': gl_object}, destfile)
 
     os.makedirs(DEST_DIRS['checks'], exist_ok=True)
     if build_all or STATIC_FILES['all_checks'] in targets:
-        allchecks = Check.template_object_all(LANG)
-        write_rst(templates['allchecks_text'], {'allchecks': allchecks}, STATIC_FILES['all_checks'])
+        allchecks = Check.template_object_all(lang)
+        templates['allchecks_text'].write_rst({'allchecks': allchecks}, STATIC_FILES['all_checks'])
 
     for tool in CheckTool.list_all():
         filename = f'examples-{tool.id}.rst'
         destfile = os.path.join(DEST_DIRS['checks'], filename)
         if build_all or destfile:
-            write_rst(templates['tool_example'], tool.example_template_object(LANG), destfile)
+            templates['tool_example'].write_rst(tool.example_template_object(lang), destfile)
 
     os.makedirs(DEST_DIRS['faq_articles'], exist_ok=True)
     for faq in Faq.list_all():
         filename = f'{faq.id}.rst'
         destfile = os.path.join(DEST_DIRS['faq_articles'], filename)
         if build_all or destfile in targets:
-            write_rst(templates['faq_article'], faq.template_object(LANG), destfile)
+            templates['faq_article'].write_rst(faq.template_object(lang), destfile)
 
     os.makedirs(DEST_DIRS['faq_tags'], exist_ok=True)
     for tag in FaqTag.list_all():
         filename = f'{tag.id}.rst'
         destfile = os.path.join(DEST_DIRS['faq_tags'], filename)
         if tag.article_count() > 0 and build_all or destfile in targets:
-            write_rst(templates['faq_tagpage'], tag.template_object(LANG), destfile)
+            templates['faq_tagpage'].write_rst(tag.template_object(lang), destfile)
 
-    sorted_tags = sorted(FaqTag.list_all(), key=lambda x: x.names[LANG])
-    tags = [tag.template_object(LANG) for tag in sorted_tags if tag.article_count() > 0]
-    tagpages = [tagpage.template_object(LANG) for tagpage in sorted_tags if tagpage.article_count() > 0]
+    sorted_tags = sorted(FaqTag.list_all(), key=lambda x: x.names[lang])
+    tags = [tag.template_object(lang) for tag in sorted_tags if tag.article_count() > 0]
+    tagpages = [tagpage.template_object(lang) for tagpage in sorted_tags if tagpage.article_count() > 0]
     if build_all or STATIC_FILES['faq_index'] in targets:
-        articles = [article.template_object(LANG) for article in Faq.list_all(sort_by='date')]
-        write_rst(templates['faq_index'], {'articles': articles, 'tags': tags}, STATIC_FILES['faq_index'])
+        articles = [article.template_object(lang) for article in Faq.list_all(sort_by='date')]
+        templates['faq_index'].write_rst({'articles': articles, 'tags': tags}, STATIC_FILES['faq_index'])
 
     if build_all or STATIC_FILES['faq_tag_index'] in targets:
-        write_rst(templates['faq_tag_index'], {'tags': tagpages}, STATIC_FILES['faq_tag_index'])
+        templates['faq_tag_index'].write_rst({'tags': tagpages}, STATIC_FILES['faq_tag_index'])
 
     if build_all or STATIC_FILES['FAQ_ARTICLE_INDEX_PATH'] in targets:
-        articles = [article.template_object(LANG) for article in Faq.list_all(sort_by='sortKey')]
-        write_rst(templates['faq_article_index'], {'articles': articles}, STATIC_FILES['faq_article_index'])
+        articles = [article.template_object(lang) for article in Faq.list_all(sort_by='sortKey')]
+        templates['faq_article_index'].write_rst({'articles': articles}, STATIC_FILES['faq_article_index'])
 
     os.makedirs(DEST_DIRS['info2gl'], exist_ok=True)
     for info in InfoRef.get_all_internals():
@@ -175,8 +145,8 @@ def main():
             destfile = os.path.join(DEST_DIRS['info2gl'], filename)
             if build_all or destfile in targets:
                 sorted_guidelines = sorted(rel.get_info_to_guidelines(info), key=lambda item: item.sort_key)
-                guidelines = [guideline.get_category_and_id(LANG) for guideline in sorted_guidelines]
-                write_rst(templates['info_to_gl'], {'guidelines': guidelines}, destfile)
+                guidelines = [guideline.get_category_and_id(lang) for guideline in sorted_guidelines]
+                templates['info_to_gl'].write_rst({'guidelines': guidelines}, destfile)
 
     os.makedirs(DEST_DIRS['info2faq'], exist_ok=True)
     for info in InfoRef.get_all_internals():
@@ -185,16 +155,16 @@ def main():
             destfile = os.path.join(DEST_DIRS['info2faq'], filename)
             makefile_vars_list['info_to_faq_target'].append(destfile)
             if build_all or destfile in targets:
-                write_rst(templates['info_to_faq'], {'faqs': [faq.id for faq in rel.get_info_to_faqs(info)]}, destfile)
+                templates['info_to_faq'].write_rst({'faqs': [faq.id for faq in rel.get_info_to_faqs(info)]}, destfile)
 
     os.makedirs(DEST_DIRS['misc'], exist_ok=True)
     if build_all or STATIC_FILES['wcag21mapping'] in targets:
-        sc_mapping = [sc.template_object(LANG) for sc in WcagSc.get_all().values() if len(rel.get_sc_to_guidelines(sc)) > 0]
-        write_rst(templates['wcag21mapping'], {'mapping': sc_mapping}, STATIC_FILES['wcag21mapping'])
+        sc_mapping = [sc.template_object(lang) for sc in WcagSc.get_all().values() if len(rel.get_sc_to_guidelines(sc)) > 0]
+        templates['wcag21mapping'].write_rst({'mapping': sc_mapping}, STATIC_FILES['wcag21mapping'])
 
     if build_all or STATIC_FILES['priority_diff'] in targets:
-        diffs = [sc.template_object(LANG) for sc in WcagSc.get_all().values() if sc.level != sc.local_priority]
-        write_rst(templates['priority_diff'], {'diffs': diffs}, STATIC_FILES['priority_diff'])
+        diffs = [sc.template_object(lang) for sc in WcagSc.get_all().values() if sc.level != sc.local_priority]
+        templates['priority_diff'].write_rst({'diffs': diffs}, STATIC_FILES['priority_diff'])
 
     if build_all or STATIC_FILES['miscdefs'] in targets:
         try:
@@ -207,10 +177,10 @@ def main():
         for link in info_links:
             external_info_links.append({
                 'label': link,
-                'text': info_links[link]['text'][LANG],
-                'url': info_links[link]['url'][LANG]
+                'text': info_links[link]['text'][lang],
+                'url': info_links[link]['url'][lang]
             })
-        write_rst(templates['miscdefs'], {'links': external_info_links}, STATIC_FILES['miscdefs'])
+        templates['miscdefs'].write_rst({'links': external_info_links}, STATIC_FILES['miscdefs'])
 
     if build_all or STATIC_FILES['makefile'] in targets:
         build_depends = []
@@ -257,8 +227,7 @@ def main():
             makefile_vars[key] = ' '.join(value)
             makefile_vars['depends'] = build_depends
         destfile = STATIC_FILES['makefile']
-        write_rst(templates['makefile'], makefile_vars, destfile)
-
+        templates['makefile'].write_rst({**makefile_vars, **MAKEFILE_VARS}, destfile)
 
 def ls_dir(dirname):
     files = []
@@ -346,39 +315,6 @@ def process_yaml_file(file_path, schema_dir, schema_file, no_check, resolver):
     data['src_path'] = os.path.relpath(file_path, start=os.getcwd())
     return data
 
-def make_heading(title, level, className=""):
-
-    def _isMultiByte(c):
-        return unicodedata.east_asian_width(c) in ['F', 'W', 'A']
-
-    def _width(c):
-        return 2 if _isMultiByte(c) else 1
-
-    def width(s):
-        return sum(_width(c) for c in s)
-
-    # Modify heading_styles accordingly
-    heading_styles = [('#', True), ('*', True), ('=', False), ('-', False), ('^', False), ('"', False)]
-
-    if not 1 <= level <= len(heading_styles):
-        raise ValueError(f'Invalid level: {level}. Must be between 1 and {len(heading_styles)}')
-
-    char, overline = heading_styles[level - 1]
-    line = char * width(title)
-
-    heading_lines = []
-
-    if className:
-        heading_lines.append(f'.. rst-class:: {className}\n')
-
-    if overline:
-        heading_lines.append(line)
-
-    heading_lines.append(title)
-    heading_lines.append(line)
-
-    return '\n'.join(heading_lines)
-
 def check_duplicate_values(lst, key, dataset):
     # Extract the values for the given key across the list of dictionaries
     values = [d[key] for d in lst]
@@ -393,77 +329,7 @@ def check_duplicate_values(lst, key, dataset):
     if non_unique_values:
         raise ValueError(f"Duplicate values in {dataset}: {non_unique_values}")
 
-def parse_args():
-    parser = argparse.ArgumentParser(description="Process YAML files into rst files for the a11y-guidelines.")
-    parser.add_argument('--no-check', action='store_true', help='Do not run various checks of YAML files')
-    parser.add_argument('--lang', '-l', type=str, choices=AVAILABLE_LANGUAGES, default='ja', help=f'the language of the output file ({" ".join(AVAILABLE_LANGUAGES)})')
-    parser.add_argument('files', nargs='*', help='Filenames')
-    return parser.parse_args()
 
-def process_arguments(args):
-    """
-    Process the command-line arguments to determine the build mode, target files, and other options.
-
-    Args:
-        args: The parsed command-line arguments.
-
-    Returns:
-        A dictionary containing settings derived from the command-line arguments.
-    """
-    settings = {
-        'build_all': not args.files,
-        'targets': args.files if args.files else [],
-        'lang': args.lang
-    }
-
-    # 例：新しいオプション 'verbose' の処理
-    # if hasattr(args, 'verbose'):
-    #     settings['verbose'] = args.verbose
-
-    return settings
-
-def setup_template_environment():
-    """
-    Set up the Jinja2 template environment.
-
-    Returns:
-        The configured Jinja2 environment object.
-    """
-    template_env = Environment(
-        loader=FileSystemLoader(TEMPLATE_DIR)
-    )
-    template_env.filters['make_heading'] = make_heading
-    return template_env
-
-def load_templates(template_env):
-    """
-    Load the necessary templates using the provided Jinja2 environment.
-
-    Args:
-        template_env: The Jinja2 environment object.
-
-    Returns:
-        A dictionary of Jinja2 templates.
-    """
-
-    templates = {name: template_env.get_template(filename) for name, filename in TEMPLATE_FILENAMES.items()}
-    return templates
-
-def write_rst(template, data, output_path):
-    """
-    Render a Jinja2 template with provided data and write the output to an RST file.
-
-    Args:
-        template: Jinja2 template object.
-        data: Data to be used in the template.
-        output_path: Path to the output RST file.
-
-    Returns:
-        None
-    """
-    rendered_content = template.render(data)
-    with open(output_path, mode='w', encoding='utf-8', newline='\n') as file:
-        file.write(rendered_content)
 
 if __name__ == "__main__":
     main()
